Add number of tokens to tokenizer
The tokenizer now supplies the number of remaining tokens to the
user.

The count is initialized by iterating to the entire string in the setup
function. This approach was choosen because the strtok best knows what
counts as a token and what doesn't, making this implementation vastly
simpler than iterating over all chars and reimplementing strtok's rules,
skipping leading and trailing delimiters.