Make aoc_main multipart compatible
The AoC days have two related requirements. This extension of aoc_main
makes it possible to handle both in the same executable.

This is made possible by now providing two functions to the aoc_main.
The first CLI parameter now denotes which part of the day should be
executed, part 1 or part 2.

Numbers 1 and 2 are the only possible inputs for this parameter.

Furthermore, aoc_main has been split into the original aoc_main, which
now only handles input verification, and exec_dayfunc, which handles the
actual execution.