Common: Update Checking functions
The check_header function did not work as intended, because the variable
with the checking result is stored in the Cache. Due to the fact that
the variable name is always HAS_HEADER, regardless of the headers name,
the test is only done once, for the first header checked. Afterwards,
any additional call of check_header checks whether HAS_HEADER is already
set and quits without checking for the new header.

The result variable's name now has the checked header's name as part of it.

Furthermore, a function similar to check_header has been added for
checking symbols.